update delete book wrong id  to passe test

+++ LibraryBackend/Controllers/BookController.cs
@@ -67,6 +67,7 @@ namespace LibraryBackend.Controllers
     [ProducesResponseType(StatusCodes.Status404NotFound)]
     public async Task<ActionResult> DeleteBook(int id)
     {
+      if(await _bookRepository.GetBookByIdAsync(id)== null) return NotFound();
       await _bookRepository.DeleteBook(id);
       return NoContent();
     }
